requirements.txt
amqp==2.4.2
argh==0.26.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-celery-beat==1.4.0
django-rest-framework==0.1.0
django-timezone-field==3.0
djangorestframework==3.9.3
idna==2.8
kombu==4.5.0
pathtools==0.1.2
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
vine==1.3.0
watchdog==0.9.0
